# labassist-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับระบบจัดการ TA ภาควิชา CS มหาวิทยาลัยศิลปากร",
        "title": "LabAssist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/instructor/applications/bulk-review": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instructor"
                ],
                "summary": "ตรวจสอบใบสมัครแบบกลุ่ม",
                "parameters": [
                    {
                        "description": "รายการใบสมัครและผลการตรวจสอบ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BulkReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BulkReviewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instructor/applications/{id}/review": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instructor"
                ],
                "summary": "ตรวจสอบใบสมัคร (อาจารย์/สตาฟ/แอดมิน)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ผลการตรวจสอบ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/student/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "รายการใบสมัครของนักศึกษา",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Application"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "สมัครเป็น TA หรือ Lab Boy",
                "parameters": [
                    {
                        "description": "ข้อมูลการสมัคร",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/student/applications/{id}/withdraw": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "ถอนใบสมัคร",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/student/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "หน้าหลักนักศึกษา",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.StudentDashboardResponse"
                        }
                    }
                }
            }
        },
        "/student/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "ดูโปรไฟล์นักศึกษา",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "แก้ไขโปรไฟล์นักศึกษา",
                "parameters": [
                    {
                        "description": "ข้อมูลโปรไฟล์ที่ต้องการแก้ไข",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.ApplyRequest": {
            "type": "object",
            "required": [
                "course_id",
                "role_applied"
            ],
            "properties": {
                "course_id": {
                    "type": "integer",
                    "example": 1
                },
                "motivation": {
                    "type": "string",
                    "example": "ต้องการช่วยสอนนักศึกษา"
                },
                "role_applied": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RoleApplied"
                        }
                    ],
                    "example": "ta"
                }
            }
        },
        "handlers.BulkReviewRequest": {
            "type": "object",
            "required": [
                "application_ids",
                "status"
            ],
            "properties": {
                "application_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "note": {
                    "type": "string"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.AppStatus"
                        }
                    ],
                    "example": "accepted"
                }
            }
        },
        "handlers.BulkReviewResponse": {
            "type": "object",
            "properties": {
                "updated": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "handlers.DashboardStats": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "integer",
                    "example": 2
                },
                "open_courses": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "handlers.ReviewRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "note": {
                    "type": "string",
                    "example": "ผ่านการคัดเลือก"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.AppStatus"
                        }
                    ],
                    "example": "accepted"
                }
            }
        },
        "handlers.StudentDashboardResponse": {
            "type": "object",
            "properties": {
                "recent_applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Application"
                    }
                },
                "recent_courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Course"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/handlers.DashboardStats"
                }
            }
        },
...
        "handlers.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "faculty": {
                    "type": "string",
                    "example": "วิทยาการคอมพิวเตอร์"
                },
                "full_name": {
                    "type": "string",
                    "example": "สมชาย ใจดี"
                },
                "year": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "models.AppStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "rejected",
                "withdrawn"
            ],
            "x-enum-varnames": [
                "AppPending",
                "AppAccepted",
                "AppRejected",
                "AppWithdrawn"
            ]
        },
        "models.Application": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "course_code": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "course_title": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "motivation": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "reviewed_at": {
                    "type": "string"
                },
                "reviewed_by_id": {
                    "type": "integer"
                },
                "reviewed_by_name": {
                    "type": "string"
                },
                "role_applied": {
                    "$ref": "#/definitions/models.RoleApplied"
                },
                "status": {
                    "$ref": "#/definitions/models.AppStatus"
                },
                "student_code": {
                    "type": "string"
                },
                "student_email": {
                    "type": "string"
                },
                "student_faculty": {
                    "type": "string"
                },
                "student_gpa": {
                    "type": "number"
                },
                "student_id": {
                    "type": "integer"
                },
                "student_name": {
                    "description": "Computed fields (not in DB)",
                    "type": "string"
                },
                "student_year": {
                    "type": "integer"
                }
            }
        },
        "models.Course": {
            "type": "object",
            "properties": {
                "academic_year": {
                    "type": "integer"
                },
                "applicant_count": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructor_id": {
                    "type": "integer"
                },
                "instructor_name": {
                    "type": "string"
                },
                "labboy_accepted": {
                    "type": "integer"
                },
                "labboy_slots": {
                    "type": "integer"
                },
                "requirements": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.CourseStatus"
                },
                "ta_accepted": {
                    "type": "integer"
                },
                "ta_slots": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CourseStatus": {
            "type": "string",
            "enum": [
                "open",
                "closing_soon",
                "closed",
                "draft"
            ],
            "x-enum-varnames": [
                "StatusOpen",
                "StatusClosingSoon",
                "StatusClosed",
                "StatusDraft"
            ]
        },
        "models.RoleApplied": {
            "type": "string",
            "enum": [
                "ta",
                "labboy"
            ],
            "x-enum-varnames": [
                "RoleTA",
                "RoleLabBoy"
            ]
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "faculty": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "google_sub": {
                    "type": "string"
                },
                "gpa": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "student_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "student",
                "instructor",
                "staff",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleStudent",
                "RoleInstructor",
                "RoleStaff",
                "RoleAdmin"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}